deployment/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-multipart
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

